+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.4.0
+
+* Added `Uris` class to contain various utility functions for inspecting and filtering uris
+* Restrict `href` in `a` and `cite` in `blockquote` and `q` to valid uris with the schemes blank, `http` or `https`
+* Only add `rel="nofollow"` for external references
+* Restrict `src` in `img` to valid uris with the schemes blank, `http`, `https` or `data`
+
+
 ## 0.3.0
 
 * Replaced `AddAttribute` by `AttributeCollector` for more flexibility in the generated attributes